pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "yaml-resume-builder"
version = "0.1.3"
description = "A package to generate PDF resumes from YAML files using LaTeX templates"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=7.1.2",
    "jinja2>=3.0.0",
    "pyyaml>=5.1",
]

[project.scripts]
yaml-resume-builder = "yaml_resume_builder.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-cov>=4.1.0",
    "ruff>=0.3.0",
    "isort>=5.12.0",
    "black>=24.1.0",
    "mypy>=1.8.0",
    "types-PyYAML>=6.0.0",
    "pytest-mypy-plugins>=3.0.0",
    "bandit>=1.7.6",
    "safety>=2.3.5",
    "pre-commit>=3.5.0",
]

[tool.setuptools]
packages = ["yaml_resume_builder"]

[tool.setuptools.package-data]
yaml_resume_builder = ["*.tex", "*.yml"]

[tool.ruff]
target-version = "py38"  # Updated to match minimum Python version
line-length = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "W",   # pycodestyle warnings
    "C90", # mccabe complexity
    "N",   # pep8-naming
    "B",   # flake8-bugbear
    "S",   # flake8-bandit
    "A",   # flake8-builtins
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "S101",  # Use of assert detected (allowed in tests)
    "B904",  # Within an except clause, raise exceptions with raise ... from err
    "S603",  # subprocess call: check for execution of untrusted input
    "S607",  # Starting a process with a partial executable path
    "B017",  # pytest.raises() should be used as a context manager
    "A002",  # Function argument shadows a Python builtin
]
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint.isort]
known-first-party = ["yaml_resume_builder"]
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.black]
line-length = 100
target-version = ["py36", "py37", "py38", "py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
# Allow untyped decorators for third-party libraries like Click
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "click.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["yaml_resume_builder/tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "--cov=yaml_resume_builder --cov-report=term-missing"
